+++ packages/web/src/lib/api.ts
@@ -3,7 +3,7 @@ import algosdk, { Algodv2, ALGORAND_MIN_TX_FEE, SuggestedParams } from "algosdk"
 import { ChainType, TxParams } from "../types";
 
 export function algoexplorerAlgod(chain: string): algosdk.Algodv2 {
-	const mainnetAlgoExplorer = new Algodv2("", "https://algoexplorerapi.io", "");
+	const mainnetAlgoExplorer = new Algodv2("", "https://node.algoexplorerapi.io", "");
 	const testnetAlgoExplorer = new Algodv2("", "https://node.testnet.algoexplorerapi.io", "");
 	switch (chain) {
 		case ChainType.MainNet: